pyproject.toml
[project]
name = "ll"
version = "0.1.0"
description = ""
authors = [
    {name = "Tanguy SALMON",email = "tanguy.salmon@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "pandas (>=2.3.3,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "psutil (>=7.1.2,<8.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
